+++ metrics-build/metrics_builder.py
@@ -430,87 +430,6 @@ def add_text_metrics(metrics_writer: ResimMetricsWriter, flight_data: dict) -> N
     )
 
 
-def main():
-    """Entry point for the metrics reader script."""
-    print("Starting metrics reader...")
-    if BATCH_METRICS_CONFIG_PATH.exists():
-        print("Running batch metrics...")
-        run_batch_metrics()
-    else:
-        print("Running test metrics...")
-        run_test_metrics()
-
-
-def run_test_metrics():
-    """Process and generate metrics for a single test run."""
-    try:
-        # Read flight data
-        flight_data = read_flight_data("processed_flight_log.json")
-
-        # Create metrics writer with a unique ID for test metrics
-        metrics_writer = ResimMetricsWriter(uuid.uuid4())
-
-        # Keep track of metric names we've added
-        added_metrics = set()
-
-        def log_metric_addition(name: str) -> None:
-            if name in added_metrics:
-                print(f"WARNING: Metric name '{name}' is being added again!")
-            else:
-                print(f"Adding metric: {name}")
-                added_metrics.add(name)
-
-        # Add all types of metrics
-        log_metric_addition("Maximum Speed")
-        add_scalar_metrics(metrics_writer, flight_data)
-
-        log_metric_addition("Speed Over Time")
-        add_speed_over_time_plot(metrics_writer, flight_data)
-
-        log_metric_addition("Altitude Over Time")
-        add_altitude_warning_plot(metrics_writer, flight_data)
-
-        log_metric_addition("Flight States Over Time")
-        add_states_over_time_plot(metrics_writer, flight_data)
-
-        log_metric_addition("X Position Over Time")
-        add_position_over_time_plot(metrics_writer, flight_data)
-
-        log_metric_addition("Speed Distribution")
-        add_speed_distribution_plot(metrics_writer, flight_data)
-
-        log_metric_addition("3D Flight Path")
-        add_plotly_metrics_from_flight_data(metrics_writer, flight_data)
-
-        log_metric_addition("Flight Summary")
-        add_text_metrics(metrics_writer, flight_data)
-
-        # Write and validate metrics
-        print("\nValidating metrics...")
-        metrics_proto = metrics_writer.write()
-
-        # Debug: Print all metric names in the proto
-        print("\nActual metrics in proto:")
-        for metric in metrics_proto.metrics_msg.metrics_data:
-            print(f"Proto metric name: '{metric.name}'")
-
-        validate_job_metrics(metrics_proto.metrics_msg)
-
-        # Write to file
-        output_path = Path("/tmp/resim/outputs/metrics.binproto")
-        output_path.parent.mkdir(parents=True, exist_ok=True)
-
-        with output_path.open("wb") as metrics_out:
-            metrics_out.write(metrics_proto.metrics_msg.SerializeToString())
-        print(f"Wrote metrics to {output_path}")
-
-    except Exception as e:
-        print(f"Error details: {str(e)}")  # Add detailed error logging
-        raise RuntimeError("Error processing metrics") from e
-
-    print("Completed processing metrics. Exiting.")
-
-
 def add_altitude_comparison_plot(metrics_writer: ResimMetricsWriter, job_to_metrics: dict) -> None:
     """Create a plot comparing altitude over time across different flights."""
     # Create Plotly figure
@@ -567,154 +486,6 @@ def add_altitude_comparison_plot(metrics_writer: ResimMetricsWriter, job_to_metr
         .with_status(MetricStatus.PASSED_METRIC_STATUS)
         .with_plotly_data(str(fig.to_json()))
     )
-
-
-def run_batch_metrics():
-    """Process and generate aggregate metrics for a batch of tests."""
-    try:
-        # Read batch config
-        with open("/tmp/resim/inputs/batch_metrics_config.json", "r", encoding="utf-8") as metrics_config_file:
-            metrics_config = json.load(metrics_config_file)
-
-        token = metrics_config["authToken"]
-        api_url = metrics_config["apiURL"]
-        batch_id = metrics_config["batchID"]
-        project_id = metrics_config["projectID"]
-
-        # Fetch all metrics for the batch (returns a dict: job_id -> metrics proto object)
-        job_to_metrics = fetch_job_metrics_by_batch(
-            token=token,
-            api_url=api_url,
-            project_id=project_id,
-            batch_id=batch_id,
-        )
-
-        max_speeds = []
-        error_counts = []
-        warning_counts = []
-        success_counts = []
-
-        for test_id, metrics_proto in job_to_metrics.items():
-            try:
-                # Example: Extract max speed from scalar metrics
-                max_speed = None
-                for metric in metrics_proto.metrics:
-                    if metric.name == "Maximum Speed" and isinstance(metric, ScalarMetric):
-                        max_speed = metric.value
-                        break
-
-                if max_speed is not None:
-                    max_speeds.append(max_speed)
-
-                # Count status occurrences
-                error_count = sum(
-                    1 for metric in metrics_proto.metrics
-                    if getattr(metric, "status", None) == MetricStatus.FAIL_BLOCK_METRIC_STATUS
-                )
-                warning_count = sum(
-                    1 for metric in metrics_proto.metrics
-                    if getattr(metric, "status", None) == MetricStatus.FAIL_WARN_METRIC_STATUS
-                )
-                success_count = sum(
-                    1 for metric in metrics_proto.metrics
-                    if getattr(metric, "status", None) == MetricStatus.PASSED_METRIC_STATUS
-                )
-
-                error_counts.append(error_count)
-                warning_counts.append(warning_count)
-                success_counts.append(success_count)
-
-            except Exception as e:
-                print(f"Error processing test {test_id}: {str(e)}")
-                continue
-
-        # Calculate batch-level metrics
-        if max_speeds:
-            avg_max_speed = sum(max_speeds) / len(max_speeds)
-            highest_speed = max(max_speeds)
-
-            total_errors = sum(error_counts)
-            total_warnings = sum(warning_counts)
-            total_successes = sum(success_counts)
-            total_samples = total_errors + total_warnings + total_successes
-
-            success_rate = (
-                (total_successes / total_samples) * 100 if total_samples > 0 else 0
-            )
-
-            metrics_writer = ResimMetricsWriter(uuid.uuid4())
-
-            # Add batch-level scalar metrics
-            (
-                metrics_writer.add_scalar_metric("Highest Recorded Speed")
-                .with_description("Maximum speed achieved across all flights")
-                .with_importance(MetricImportance.HIGH_IMPORTANCE)
-                .with_value(highest_speed)
-                .with_unit("m/s")
-                .with_status(MetricStatus.PASSED_METRIC_STATUS)
-            )
-
-            (
-                metrics_writer.add_scalar_metric("Average Max Speed")
-                .with_description("Average of maximum speeds across all flights")
-                .with_importance(MetricImportance.HIGH_IMPORTANCE)
-                .with_value(avg_max_speed)
-                .with_unit("m/s")
-                .with_status(MetricStatus.PASSED_METRIC_STATUS)
-            )
-
-            (
-                metrics_writer.add_scalar_metric("Overall Success Rate")
-                .with_description("Percentage of successful flight states")
-                .with_importance(MetricImportance.HIGH_IMPORTANCE)
-                .with_value(success_rate)
-                .with_unit("%")
-                .with_status(
-                    MetricStatus.FAIL_BLOCK_METRIC_STATUS
-                    if success_rate < 70
-                    else (
-                        MetricStatus.FAIL_WARN_METRIC_STATUS
-                        if success_rate < 74
-                        else MetricStatus.PASSED_METRIC_STATUS
-                    )
-                )
-            )
-
-            # Add the altitude comparison plot
-            add_altitude_comparison_plot(metrics_writer, job_to_metrics)
-
-            # Add a summary text metric
-            summary = f"""# Flight Batch Summary
-- Total Flights: {len(max_speeds)}
-- Highest Speed: {highest_speed:.2f} m/s
-- Average Max Speed: {avg_max_speed:.2f} m/s
-- Total Errors: {total_errors}
-- Total Warnings: {total_warnings}
-- Success Rate: {success_rate:.1f}%
-"""
-
-            (
-                metrics_writer.add_text_metric("Batch Summary")
-                .with_description("Summary of all flight data")
-                .with_importance(MetricImportance.HIGH_IMPORTANCE)
-                .with_status(MetricStatus.PASSED_METRIC_STATUS)
-                .with_text(summary)
-            )
-
-            # Write and validate metrics
-            metrics_proto = metrics_writer.write()
-            validate_job_metrics(metrics_proto.metrics_msg)
-
-            # Write to file
-            output_path = Path("/tmp/resim/outputs/metrics.binproto")
-            with output_path.open("wb") as metrics_out:
-                metrics_out.write(metrics_proto.metrics_msg.SerializeToString())
-            print(f"Batch metrics: Wrote metrics to {output_path}")
-
-    except Exception as e:
-        raise RuntimeError("Error processing batch metrics") from e
-
-    print("Completed processing batch metrics. Exiting.")
 
 
 def add_plotly_metrics_from_proto(metrics_writer: ResimMetricsWriter, metrics_proto, job_id=None, job_name=None) -> None:
@@ -781,8 +552,4 @@ def add_plotly_metrics_from_proto(metrics_writer: ResimMetricsWriter, metrics_pr
         .with_importance(MetricImportance.HIGH_IMPORTANCE)
         .with_status(MetricStatus.PASSED_METRIC_STATUS)
         .with_plotly_data(str(fig.to_json()))
-    )
-
-
-if __name__ == "__main__":
-    main()
+    )
